Change signature of GradientCollection set
The set method of GradientCollection trait was expecting a reference
to the gradients to be set.

However, this method is only used to update the gradients, which have
been computed somewhere else. For the gradients we don't want to
store references, but the actual owned data. The getter can be used
to retrieve *references* to the gradients though.