Add Matrix, Matrix implementation for optimizer state and GD
This should be a temporary measure, while I think of something
more elegant. The problem is that I defined my SingleLayerClassifier
as a model that inputs a matrix and outputs a matrix, not a vector
like the other models so far.

GradientDescent and its OptimizerState were implemented for Matrix,
Vector, so they didn't work with the SingleLayerClassifier, so I
just copied the implementations and changed the Output type from
Vector to Matrix.

This is not ideal because the code is exactly the same, but for now
I'll use this as a quick and dirty solution, because it's late and
I want to go to bed.

a